[step2] refactor[ladder] Positions 일급콜렉션 사용하도록 수정
